+++ tests/integ/scala/test_dataframe_suite.py
@@ -41,6 +41,7 @@ from snowflake.snowpark.types import (
     DecimalType,
     DoubleType,
     FloatType,
+    GeographyType,
     IntegerType,
     LongType,
     MapType,
@@ -687,6 +688,51 @@ def test_sample_with_frac(session):
     assert len(df.sample(frac=1.0).collect()) == row_count
 
 
+def test_sample_with_seed(session):
+    row_count = 10000
+    temp_table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.range(row_count).write.save_as_table(
+        temp_table_name, create_temp_table=True
+    )
+    df = session.table(temp_table_name)
+    try:
+        sample1 = df.sample(frac=0.1, seed=1).collect()
+        sample2 = df.sample(frac=0.1, seed=1).collect()
+        Utils.check_answer(sample1, sample2, sort=True)
+    finally:
+        Utils.drop_table(session, temp_table_name)
+
+
+def test_sample_with_sampling_method(session):
+    """sampling method actually has no impact on result. It has impact on performance."""
+    row_count = 10000
+    temp_table_name = Utils.random_name_for_temp_object(TempObjectType.TABLE)
+    session.range(row_count).write.save_as_table(
+        temp_table_name, create_temp_table=True
+    )
+    df = session.table(temp_table_name)
+    try:
+        assert df.sample(frac=0.0, sampling_method="BLOCK").count() == 0
+        half_row_count = row_count * 0.5
+        assert (
+            abs(df.sample(frac=0.5, sampling_method="BLOCK").count() - half_row_count)
+            < half_row_count * SAMPLING_DEVIATION
+        )
+        assert df.sample(frac=1.0, sampling_method="BLOCK").count() == row_count
+        assert len(df.sample(frac=0.0, sampling_method="BLOCK").collect()) == 0
+        half_row_count = row_count * 0.5
+        assert (
+            abs(
+                len(df.sample(frac=0.5, sampling_method="BLOCK").collect())
+                - half_row_count
+            )
+            < half_row_count * SAMPLING_DEVIATION
+        )
+        assert len(df.sample(frac=1.0, sampling_method="BLOCK").collect()) == row_count
+    finally:
+        Utils.drop_table(session, temp_table_name)
+
+
 def test_sample_negative(session):
     """Tests negative test cases for sample"""
     row_count = 10000
@@ -699,6 +745,10 @@ def test_sample_negative(session):
         df.sample(frac=-0.01)
     with pytest.raises(ValueError):
         df.sample(frac=1.01)
+
+    table = session.table("non_existing_table")
+    with pytest.raises(ValueError):
+        table.sample(sampling_method="InvalidValue")
 
 
 def test_sample_on_join(session):
@@ -1431,21 +1481,32 @@ def test_createDataFrame_with_given_schema_array_map_variant(session):
             StructField("array", ArrayType(None)),
             StructField("map", MapType(None, None)),
             StructField("variant", VariantType()),
-            # StructField("geography", GeographyType()),
+            StructField("geography", GeographyType()),
         ]
     )
-    data = [Row(["'", 2], {"'": 1}, 1), Row(None, None, None)]
+    data = [
+        Row(["'", 2], {"'": 1}, 1, "POINT(30 10)"),
+        Row(None, None, None, None),
+    ]
     df = session.create_dataframe(data, schema)
     assert (
         str(df.schema)
         == "StructType[StructField(ARRAY, ArrayType[String], Nullable=True), "
         "StructField(MAP, MapType[String, String], Nullable=True), "
-        "StructField(VARIANT, Variant, Nullable=True)]"
+        "StructField(VARIANT, Variant, Nullable=True), "
+        "StructField(GEOGRAPHY, GeographyType, Nullable=True)]"
     )
     df.show()
+    geography_string = """{
+  "coordinates": [
+    30,
+    10
+  ],
+  "type": "Point"
+}"""
     expected = [
-        Row('[\n  "\'",\n  2\n]', '{\n  "\'": 1\n}', "1"),
-        Row(None, None, None),
+        Row('[\n  "\'",\n  2\n]', '{\n  "\'": 1\n}', "1", geography_string),
+        Row(None, None, None, None),
     ]
     Utils.check_answer(df, expected, sort=False)