+++ tests/integ/scala/test_function_suite.py
@@ -129,6 +129,7 @@ from snowflake.snowpark.functions import (
     sum_distinct,
     to_array,
     to_date,
+    to_geography,
     to_json,
     to_object,
     to_timestamp,
@@ -2318,6 +2319,26 @@ def test_to_xml(session):
     )
 
 
+def test_to_geography(session):
+    geography_string = """{
+  "coordinates": [
+    30,
+    10
+  ],
+  "type": "Point"
+}"""
+    geography = TestData.geography(session)
+    Utils.check_answer(
+        geography.select(to_geography(col("a"))),
+        [Row(geography_string)],
+        sort=False,
+    )
+    assert geography.select(to_geography(col("a"))).collect()[0][0] == geography_string
+
+    # same as above, but pass str instead of Column
+    assert geography.select(to_geography("a")).collect()[0][0] == geography_string
+
+
 @pytest.mark.parametrize("a", ["a", col("a")])
 def test_typeof(session, a):
     Utils.check_answer(
@@ -2521,6 +2542,13 @@ def test_iff(session):
     )
     Utils.check_answer(
         df.select("b", "c", "d", iff(col("b") == col("c"), col("b"), col("d"))),
+        [Row(2, 2, 4, 2), Row(12, 12, 14, 12), Row(22, 23, 24, 24)],
+        sort=False,
+    )
+
+    # accept sql expression
+    Utils.check_answer(
+        df.select("b", "c", "d", iff("b = c", col("b"), col("d"))),
         [Row(2, 2, 4, 2), Row(12, 12, 14, 12), Row(22, 23, 24, 24)],
         sort=False,
     )